pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "cvbench"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "tensorflow",
    "keras",
    "keras-hub",
    "click",
    "pyyaml",
    "scikit-learn",
    "matplotlib",
    "pandas",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "commitizen",
]

[project.scripts]
train    = "cli.train:train"
evaluate = "cli.evaluate:evaluate"
predict  = "cli.predict:predict"
runs     = "cli.runs:runs"
generate = "cli.generate:generate"

[tool.setuptools.packages.find]
where = ["."]
include = ["core*", "cli*", "augmentations*"]

[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "v$version"
version_scheme = "semver"
version_provider = "pep621"
update_changelog_on_bump = true
major_version_zero = true

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "tf: marks tests that require TensorFlow (deselect with '-m \"not tf\"')",
]
